+++ packages/dsw-document-worker/dsw/document_worker/__main__.py
@@ -2,5 +2,4 @@ from . import consts
 from .cli import main
 
 
-# pylint: disable-next=no-value-for-parameter
 main(prog_name=consts.PROG_NAME)