+++ packages/dsw-document-worker/dsw/document_worker/model/context.py
@@ -1,4 +1,3 @@
-# pylint: disable=too-many-lines, unused-argument, too-many-arguments,
 import abc
 import datetime
 import re
@@ -8,7 +7,7 @@ import dateutil.parser as dp
 
 from .. import consts
 from ..utils import check_metamodel_version
-from .utils import strip_markdown
+from .utils import render_markdown, strip_markdown
 
 
 AnnotationsT = dict[str, str | list[str]]
@@ -125,18 +124,14 @@ class SimpleAuthor:
 class User:
 
     def __init__(self, *, uuid: str, first_name: str, last_name: str, email: str,
-                 role: str, created_at: datetime.datetime, updated_at: datetime.datetime,
-                 affiliation: str | None, permissions: list[str], sources: list[str],
-                 image_url: str | None):
+                 created_at: datetime.datetime, updated_at: datetime.datetime,
+                 affiliation: str | None, image_url: str | None):
         self.uuid = uuid
         self.first_name = first_name
         self.last_name = last_name
         self.email = email
-        self.role = role
         self.image_url = image_url
         self.affiliation = affiliation
-        self.permissions = permissions
-        self.sources = sources
         self.created_at = created_at
         self.updated_at = updated_at
 
@@ -149,11 +144,8 @@ class User:
             first_name=data['firstName'],
             last_name=data['lastName'],
             email=data['email'],
-            role=data['role'],
             image_url=data['imageUrl'],
             affiliation=data['affiliation'],
-            permissions=data['permissions'],
-            sources=data['sources'],
             created_at=_datetime(data['createdAt']),
             updated_at=_datetime(data['updatedAt']),
         )
@@ -266,11 +258,10 @@ class ResourcePage:
 
 class Integration(abc.ABC):
 
-    def __init__(self, *, uuid: str, name: str, variables: list[str],
+    def __init__(self, *, uuid: str, name: str,
                  integration_type: str, annotations: AnnotationsT):
         self.uuid = uuid
         self.name = name
-        self.variables = variables
         self.type = integration_type
         self.annotations = annotations
 
@@ -301,10 +292,10 @@ class ApiIntegration(Integration):
         super().__init__(
             uuid=uuid,
             name=name,
-            variables=variables,
             annotations=annotations,
             integration_type='ApiIntegration',
         )
+        self.variables = variables
         self.allow_custom_reply = allow_custom_reply
         self.request_method = request_method
         self.request_url = request_url
@@ -352,108 +343,29 @@ class ApiIntegration(Integration):
         )
 
 
-class ApiLegacyIntegration(Integration):
+class PluginIntegration(Integration):
 
-    def __init__(self, *, uuid: str, name: str, integration_id: str,
-                 item_url: str | None, logo: str | None,
-                 variables: list[str], rq_body: str, rq_method: str,
-                 rq_headers: dict[str, str], rq_url: str, rs_list_field: str | None,
-                 rs_item_id: str | None, rs_item_template: str,
+    def __init__(self, *, uuid: str, name: str,
+                 plugin_uuid: str, integration_id: str, settings: dict,
                  annotations: AnnotationsT):
         super().__init__(
             uuid=uuid,
             name=name,
-            variables=variables,
             annotations=annotations,
-            integration_type='ApiLegacyIntegration',
+            integration_type='PluginIntegration',
         )
-        self.id = integration_id
-        self.item_url = item_url
-        self.logo = logo
-        self.rq_body = rq_body
-        self.rq_method = rq_method
-        self.rq_url = rq_url
-        self.rq_headers = rq_headers
-        self.rs_list_field = rs_list_field
-        self.rs_item_id = rs_item_id
-        self.rs_item_template = rs_item_template
-
-    def item(self, item_id: str) -> str | None:
-        if self.item_url is None:
-            return None
-        return self.item_url.replace('${id}', item_id)
-
-    @staticmethod
-    def default():
-        return ApiLegacyIntegration(
-            uuid=consts.NULL_UUID,
-            name='',
-            logo='',
-            integration_id='',
-            item_url='',
-            variables=[],
-            rq_body='',
-            rq_method='GET',
-            rq_url='',
-            rq_headers={},
-            rs_list_field='',
-            rs_item_id='',
-            rs_item_template='',
-            annotations={},
-        )
+        self.plugin_uuid = plugin_uuid
+        self.integration_id = integration_id
+        self.settings = settings
 
     @staticmethod
     def load(data: dict, **options):
-        return ApiLegacyIntegration(
+        return PluginIntegration(
             uuid=data['uuid'],
             name=data['name'],
-            integration_id=data['id'],
-            item_url=data['itemUrl'],
-            logo=data['logo'],
-            variables=data['variables'],
-            rq_body=data['requestBody'],
-            rq_headers=data['requestHeaders'],
-            rq_method=data['requestMethod'],
-            rq_url=data['requestUrl'],
-            rs_list_field=data['responseListField'],
-            rs_item_id=data['responseItemId'],
-            rs_item_template=data['responseItemTemplate'],
-            annotations=_load_annotations(data['annotations']),
-        )
-
-
-class WidgetIntegration(Integration):
-
-    def __init__(self, *, uuid: str, name: str, integration_id: str,
-                 item_url: str | None, logo: str | None, variables: list[str],
-                 widget_url: str, annotations: AnnotationsT):
-        super().__init__(
-            uuid=uuid,
-            name=name,
-            variables=variables,
-            annotations=annotations,
-            integration_type='WidgetIntegration',
-        )
-        self.id = integration_id
-        self.item_url = item_url
-        self.logo = logo
-        self.widget_url = widget_url
-
-    def item(self, item_id: str) -> str | None:
-        if self.item_url is None:
-            return None
-        return self.item_url.replace('${id}', item_id)
-
-    @staticmethod
-    def load(data: dict, **options):
-        return WidgetIntegration(
-            uuid=data['uuid'],
-            name=data['name'],
-            integration_id=data['id'],
-            item_url=data['itemUrl'],
-            logo=data['logo'],
-            variables=data['variables'],
-            widget_url=data['widgetUrl'],
+            plugin_uuid=data['pluginUuid'],
+            integration_id=data['pluginIntegrationId'],
+            settings=data['pluginSettings'],
             annotations=_load_annotations(data['annotations']),
         )
 
@@ -767,6 +679,14 @@ class StringReply(Reply):
         self.value = value
 
     @property
+    def markdown_html(self) -> str:
+        return render_markdown(self.value)
+
+    @property
+    def markdown_plain(self) -> str:
+        return strip_markdown(self.value)
+
+    @property
     def as_number(self) -> float | None:
         try:
             return float(self.value)
@@ -884,7 +804,7 @@ class IntegrationReply(Reply):
 
     def __init__(self, *, path: str, created_at: datetime.datetime,
                  created_by: SimpleAuthor | None, value: str, value_type: str,
-                 item_id: str | None, raw: typing.Any | None = None):
+                 raw: typing.Any | None = None):
         super().__init__(
             path=path,
             created_at=created_at,
@@ -892,34 +812,16 @@ class IntegrationReply(Reply):
             reply_type='IntegrationReply',
         )
         self.value_type = value_type
-        self.item_id = item_id
         self.raw = raw
         self.value = value
-
-    @property
-    def id(self) -> str | None:
-        return self.item_id
 
     @property
     def is_plain(self) -> bool:
         return self.value_type == 'PlainType'
 
     @property
-    def is_legacy_integration(self) -> bool:
-        return self.value_type == 'IntegrationLegacyType'
-
-    @property
     def is_integration(self) -> bool:
         return self.value_type == 'IntegrationType'
-
-    @property
-    def url(self) -> str | None:
-        if not self.is_legacy_integration or self.item_id is None:
-            return None
-        if isinstance(self.question, IntegrationQuestion) \
-                and isinstance(self.question.integration, ApiLegacyIntegration):
-            return self.question.integration.item(self.item_id)
-        return None
 
     def resolve_links(self, ctx):
         super().resolve_links_parent(ctx)
@@ -942,7 +844,6 @@ class IntegrationReply(Reply):
             created_by=SimpleAuthor.load(data['createdBy'], **options),
             value_type=data['value']['value']['type'],
             value=data['value']['value'].get('value', ''),
-            item_id=data['value']['value'].get('id'),
             raw=data['value']['value'].get('raw'),
         )
 
@@ -1622,8 +1523,7 @@ _REFERENCE_TYPES: dict[str, type[Reference]] = {
 
 _INTEGRATION_TYPES: dict[str, type[Integration]] = {
     'ApiIntegration': ApiIntegration,
-    'ApiLegacyIntegration': ApiLegacyIntegration,
-    'WidgetIntegration': WidgetIntegration,
+    'PluginIntegration': PluginIntegration,
 }
 
 
@@ -1823,11 +1723,11 @@ class ContextConfig:
 
 class Document:
 
-    def __init__(self, *, uuid: str, name: str, document_template_id: str, format_uuid: str,
+    def __init__(self, *, uuid: str, name: str, document_template_uuid: str, format_uuid: str,
                  created_by: User | None, created_at: datetime.datetime):
         self.uuid = uuid
         self.name = name
-        self.document_template_uuid = document_template_id
+        self.document_template_uuid = document_template_uuid
         self.format_uuid = format_uuid
         self.created_by = created_by
         self.created_at = created_at
@@ -1837,7 +1737,7 @@ class Document:
         return Document(
             uuid=data['uuid'],
             name=data['name'],
-            document_template_id=data['documentTemplateId'],
+            document_template_uuid=data['documentTemplateUuid'],
             format_uuid=data['formatUuid'],
             created_by=User.load(data['createdBy'], **options),
             created_at=_datetime(data['createdAt']),
@@ -1912,10 +1812,6 @@ class ProjectFile:
         self.reply: FileReply | None = None
         self.download_url: str = ''
         self.project_uuid: str | None = None
-
-    @property
-    def questionnaire_uuid(self) -> str | None:
-        return self.project_uuid
 
     def resolve_links(self, ctx):
         self.project_uuid = ctx.project.uuid
@@ -1992,7 +1888,7 @@ class Project:
         return entity
 
 
-class Package:
+class KnowledgeModelPackage:
 
     def __init__(self, *, org_id: str, km_id: str, version: str, versions: list[str],
                  name: str, description: str, created_at: datetime.datetime):
@@ -2012,7 +1908,7 @@ class Package:
 
     @staticmethod
     def load(data: dict, **options):
-        return Package(
+        return KnowledgeModelPackage(
             org_id=data['organizationId'],
             km_id=data['kmId'],
             version=data['version'],
@@ -2257,10 +2153,10 @@ class DocumentContext:
         )
         self.config = ContextConfig.load(ctx['config'], **options)
         self.km = KnowledgeModel.load(ctx['knowledgeModel'], **options)
-        self.project = Project.load(ctx['questionnaire'], **options)
+        self.project = Project.load(ctx['project'], **options)
         self.report = Report.load(ctx['report'], **options)
         self.document = Document.load(ctx['document'], **options)
-        self.package = Package.load(ctx['package'], **options)
+        self.km_package = KnowledgeModelPackage.load(ctx['knowledgeModelPackage'], **options)
         self.organization = Organization.load(ctx['organization'], **options)
         self.current_phase: Phase = PHASE_NEVER
 
@@ -2278,20 +2174,12 @@ class DocumentContext:
         return self.config
 
     @property
-    def qtn(self) -> Project:
-        return self.project
-
-    @property
-    def questionnaire(self) -> Project:
-        return self.project
-
-    @property
     def org(self) -> Organization:
         return self.organization
 
     @property
-    def pkg(self) -> Package:
-        return self.package
+    def pkg(self) -> KnowledgeModelPackage:
+        return self.km_package
 
     @property
     def doc(self) -> Document: