+++ packages/dsw-document-worker/dsw/document_worker/plugins/manager.py
@@ -4,7 +4,6 @@ from .. import consts
 
 
 def create_manager():
-    # pylint: disable-next=import-outside-toplevel
     import dsw.document_worker.plugins.specs as hookspecs
 
     pm = pluggy.PluginManager(consts.PACKAGE_NAME)