+++ packages/dsw-document-worker/dsw/document_worker/templates/extraction.py
@@ -132,18 +132,9 @@ class SimpleExtractor(ContextExtractor):
                         '_value': reply.value,
                         '_path': path,
                     }
-                elif reply.is_legacy_integration:
-                    self._objects[key][a_key] = {
-                        '_value': reply.value,
-                        '_id': reply.id,
-                        '_url': reply.url,
-                        '_path': path,
-                    }
                 else:
                     self._objects[key][a_key] = {
                         '_value': reply.value,
-                        '_id': reply.id,
-                        '_url': reply.url,
                         '_raw': reply.raw,
                         '_path': path,
                     }