+++ packages/dsw-tdk/dsw/tdk/__main__.py
@@ -1,5 +1,4 @@
 from .cli import main
 
 
-# pylint: disable-next=no-value-for-parameter
 main(ctx=None)